refactor: tighten authoring architecture
