+++ apps/desktop/src/renderer/src/components/graph/edge-select-event.ts
@@ -1,8 +1,1 @@
-import type { RefEdgeData } from './schema-to-graph';
-
 export const TYPE_EDGE_SELECT_EVENT = 'contexture:edge-select';
-
-export interface EdgeSelection {
-  edgeId: string;
-  data: RefEdgeData;
-}
